Adds voltage control mechanisms

+++ src/rate.rs
@@ -1,10 +1,10 @@
 pub trait Rate: Copy + Clone + std::fmt::Debug + Sized + Default {
-    const RECIPROCAL: u32;
+    const RECIPROCAL: i32;
 }
 
 #[derive(Debug, Clone, Copy, Default)]
 pub struct Hz44100 {}
 
 impl Rate for Hz44100 {
-    const RECIPROCAL: u32 = u32::MAX / 44_100;
+    const RECIPROCAL: i32 = (u32::MAX / 44_100) as i32;
 }
